+++ hexagonal-intro/src/test/java/com/baeldung/hexagonal/service/DomainOrderServiceTest.java
@@ -3,11 +3,8 @@ package com.baeldung.hexagonal.service;
 import com.baeldung.hexagonal.domain.CustomerType;
 import com.baeldung.hexagonal.domain.Order;
 import com.baeldung.hexagonal.repository.OrderRepository;
-import com.baeldung.hexagonal.repository.mongodb.MongoDbOrderRepository;
-import com.baeldung.hexagonal.repository.mongodb.SpringDataMongoDbOrderRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.mockito.InjectMocks;
 import org.mockito.Mock;
 
 import java.util.UUID;
@@ -18,17 +15,17 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.mockito.Mockito.mock;
 import static org.mockito.MockitoAnnotations.initMocks;
 
-public class OrderServiceTest {
+public class DomainOrderServiceTest {
 
     @Mock
     private OrderRepository springDataOrderRepository;
 
-    private OrderService orderService;
+    private DomainOrderService domainOrderService;
 
     @BeforeEach
     void setUp() {
         initMocks(this);
-        orderService = new OrderService(springDataOrderRepository);
+        domainOrderService = new DomainOrderService(springDataOrderRepository);
     }
 
     @Test
@@ -36,7 +33,7 @@ public class OrderServiceTest {
         Order o  = new Order();
         o.setAmount(100d);
         o.setCustomerType(CustomerType.EXEMPT);
-        orderService.applyVAT(o);
+        domainOrderService.applyVAT(o);
         assertEquals(0d,o.getVat());
     }
 
@@ -45,7 +42,7 @@ public class OrderServiceTest {
         Order o  = new Order();
         o.setAmount(100d);
         o.setCustomerType(CustomerType.REGULAR);
-        orderService.applyVAT(o);
+        domainOrderService.applyVAT(o);
         assertNotEquals(0d,o.getVat());
     }
 
@@ -56,7 +53,7 @@ public class OrderServiceTest {
         o.setId(id);
         o.setAmount(100d);
         o.setCustomerType(CustomerType.REGULAR);
-        orderService.applyVAT(o);
+        domainOrderService.applyVAT(o);
         assertEquals(id,o.getId());
     }
 
@@ -65,7 +62,7 @@ public class OrderServiceTest {
         Order o  = new Order();
         o.setAmount(100d);
         o.setCustomerType(CustomerType.REGULAR);
-        orderService.applyVAT(o);
+        domainOrderService.applyVAT(o);
         assertNotNull(o.getVat());
     }
 }
